+++ src/extension.ts
@@ -29,10 +29,13 @@ export async function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(vscode.workspace.registerTextDocumentContentProvider(textDocumentScheme, new ALObjectHelperDocumentProvider()));
 	context.subscriptions.push(vscode.languages.registerDefinitionProvider("al", new ALDefinitionProvider(reader)));
 	context.subscriptions.push(vscode.languages.registerHoverProvider("al", new ALHoverProvider(reader)));
-	context.subscriptions.push(vscode.window.registerTreeDataProvider("alObjectHelperObjectList", new ALObjectHelperTreeDataProvider()));
+	// context.subscriptions.push(vscode.window.registerTreeDataProvider("alObjectHelperObjectList", new ALObjectHelperTreeDataProvider()));
 
 	context.subscriptions.push(vscode.commands.registerCommand("al-object-helper.openALObject", async function () {
-		await reader.startReading();
+		if (reader.autoReloadObjects){
+			await reader.startReading();
+		}
+		
 		const alObject = await UIManagement.selectALObjectInApps(reader.alApps);
 		if (!alObject) {
 			return;
@@ -47,7 +50,10 @@ export async function activate(context: vscode.ExtensionContext) {
 	}));
 
 	context.subscriptions.push(vscode.commands.registerCommand("al-object-helper.openALObjectOfApp", async function () {
-		await reader.startReading();
+		if (reader.autoReloadObjects){
+			await reader.startReading();
+		}
+
 		const alApp = await UIManagement.selectALApp(reader.alApps);
 		if (!alApp) {
 			return;
@@ -67,6 +73,10 @@ export async function activate(context: vscode.ExtensionContext) {
 	}));
 
 	context.subscriptions.push(vscode.commands.registerCommand("al-object-helper.copyEvent", async function () {
+		if (reader.autoReloadObjects){
+			await reader.startReading();
+		}
+		
 		var alObjects: ALObject[] = [];
 		reader.alApps.forEach(alApp => {
 			alApp.alObjects.forEach(alObject => {
@@ -95,6 +105,10 @@ export async function activate(context: vscode.ExtensionContext) {
 	}));
 
 	context.subscriptions.push(vscode.commands.registerCommand("al-object-helper.jumpToEventSubscriber", async function () {
+		if (reader.autoReloadObjects){
+			await reader.startReadingLocalApps(reader.alApps.filter(alApp => alApp.appType === AppType.local));
+		}
+		
 		const alFunction = await UIManagement.selectEventSubscriber(reader.alApps.filter(alApp => alApp.appType === AppType.local));
 		if (!alFunction) {
 			return;
@@ -126,6 +140,10 @@ export async function activate(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(vscode.commands.registerCommand("al-object-helper.openALObjectList", async function () {
 		UIManagement.showObjectListPanel(reader.alApps);
+	}));
+
+	context.subscriptions.push(vscode.commands.registerCommand("al-object-helper.reloadObjects", async function () {
+		await reader.startReading();
 	}));
 
 	context.subscriptions.push(vscode.window.onDidChangeActiveTextEditor((textEditor) => {